refactor(katana-rpc): restructure modules
commit-id:57af318e
