+++ crates/katana/rpc/src/api/mod.rs
@@ -1,0 +1,2 @@
+pub mod katana;
+pub mod starknet;
