+++ crates/katana/test/test_data/cairo0_contract.json
