+++ crates/katana/test/test_data/cairo1_contract.json
